backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile backend/requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.12
    # via
    #   aiohttp-retry
    #   langchain
    #   pinecone
aiohttp-retry==2.9.1
    # via pinecone
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   pinecone
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   pip-tools
    #   rich-toolkit
    #   typer
    #   uvicorn
comm==0.2.2
    # via ipykernel
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
executing==2.2.0
    # via stack-data
fastapi==0.111.1
    # via -r backend/requirements.in
fastapi-cli==0.0.7
    # via fastapi
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.40.3
    # via google-genai
google-genai==1.20.0
    # via -r backend/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   google-genai
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
ipykernel==6.29.5
    # via -r backend/requirements.in
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via fastapi
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
langchain==0.2.10
    # via -r backend/requirements.in
langchain-core==0.2.43
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.2.4
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via ipykernel
numpy==1.26.4
    # via langchain
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   build
    #   ipykernel
    #   langchain-core
    #   pinecone-plugin-assistant
    #   pytest
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pinecone[asyncio]==7.0.2
    # via -r backend/requirements.in
pinecone-plugin-assistant==1.7.0
    # via pinecone
pinecone-plugin-interface==0.0.7
    # via pinecone
pip-tools==7.4.1
    # via -r backend/requirements.in
platformdirs==4.3.8
    # via jupyter-core
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
    #   rich
pypdf==4.2.0
    # via -r backend/requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r backend/requirements.in
    #   pytest-asyncio
pytest-asyncio==1.0.0
    # via -r backend/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pinecone
python-dotenv==1.0.1
    # via
    #   -r backend/requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
    #   uvicorn
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.4
    # via
    #   google-genai
    #   langchain
    #   langsmith
    #   pinecone-plugin-assistant
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==14.0.0
    # via
    #   -r backend/requirements.in
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
stack-data==0.6.3
    # via ipython
starlette==0.37.2
    # via fastapi
tenacity==8.5.0
    # via
    #   -r backend/requirements.in
    #   langchain
    #   langchain-core
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via -r backend/requirements.in
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   fastapi
    #   google-genai
    #   langchain-core
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via
    #   pinecone
    #   requests
uvicorn[standard]==0.30.1
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
wheel==0.45.1
    # via pip-tools
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
